+++ src/main/java/com/blackrook/rookscript/tools/ScriptExecutor.java
@@ -45,6 +45,7 @@ import com.blackrook.rookscript.functions.DigestFunctions;
 import com.blackrook.rookscript.functions.FileSystemFunctions;
 import com.blackrook.rookscript.functions.JSONFunctions;
 import com.blackrook.rookscript.functions.PrintFunctions;
+import com.blackrook.rookscript.functions.RandomFunctions;
 import com.blackrook.rookscript.functions.SystemFunctions;
 
 /**
@@ -77,6 +78,7 @@ public final class ScriptExecutor
 		new Resolver("Buffer", BufferFunctions.createResolver()),
 		new Resolver("Error", ErrorFunctions.createResolver()),
 		new Resolver("Math", MathFunctions.createResolver()),
+		new Resolver("Random", RandomFunctions.createResolver()),
 		new Resolver("RegEx", RegexFunctions.createResolver()),
 		new Resolver("Date / Time", DateFunctions.createResolver()),
 		new Resolver("File System", FileSystemFunctions.createResolver()),